% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_themer.R
\name{start_themer}
\alias{start_themer}
\title{Title}
\usage{
start_themer()
}
\value{
Starts theme selector
}
\description{
Title
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
start_themer()
\dontshow{\}) # examplesIf}
}
